Extract utility method into helper package
