+++ OTAnalytics/plugin_ui/nicegui_gui/ui_factory.py
@@ -14,10 +14,12 @@ from OTAnalytics.application.analysis.traffic_counting_specification import (
     CountingSpecificationDto,
 )
 from OTAnalytics.application.application import CancelAddFlow
-from OTAnalytics.application.logger import logger
 from OTAnalytics.application.resources.resource_manager import ResourceManager
 from OTAnalytics.application.use_cases.generate_flows import FlowNameGenerator
 from OTAnalytics.domain.geometry import RelativeOffsetCoordinate
+from OTAnalytics.helpers.file_helpers import (
+    get_all_files_with_correct_file_ending_in_directory,
+)
 from OTAnalytics.plugin_ui.nicegui_gui.dialogs.edit_flow_dialog import EditFlowDialog
 from OTAnalytics.plugin_ui.nicegui_gui.dialogs.edit_section_dialog import (
     EditSectionDialog,
@@ -110,30 +112,9 @@ class NiceGuiUiFactory(UiFactory):
             extension_options=extension_options,
             multiple=True,
         )
-        return self.get_all_files_with_correct_file_ending_in_directory(
+        return get_all_files_with_correct_file_ending_in_directory(
             file_paths, converted_filetypes
         )
-
-    def get_all_files_with_correct_file_ending_in_directory(
-        self, files: Iterable[Path], file_types: list[str]
-    ) -> list[Path]:
-        files_to_save: set[Path] = set()
-        for file_type in file_types:
-            for file in files:
-                file_to_save = file.expanduser()
-                if file_to_save.is_dir():
-                    files_in_directory = file_to_save.rglob(file_type)
-                    files_to_save.update(files_in_directory)
-                    continue
-
-                if not file_to_save.exists() or file_to_save.suffix != file_type:
-                    logger().warning(
-                        f"Ottrk file'{file_to_save}' does not exist. Skipping file."
-                    )
-                    continue
-
-                files_to_save.add(file_to_save)
-        return list(files_to_save)
 
     async def ask_for_save_file_path(
         self,
